Перешел на спринг бут версии 3.5.14, т.к. был конфликт зависимостей junit на версии 4.х.х

+++ src/main/java/ru/netology/repository/CloudServiceRepository.java
@@ -57,6 +57,15 @@ public class CloudServiceRepository {
      */
     public void saveFile(MultipartFile file, String fileName){
         String path = filesPath + "/" + fileName;
+        File dir = new File(filesPath);
+        if(!dir.exists()){
+            if(!dir.mkdir()) {
+                throw new ErrorUploadFileException();
+            }
+        }
+        if(Files.exists(Path.of(path))){
+            throw new ErrorUploadFileException();
+        }
         try(FileOutputStream fos = new FileOutputStream(path)) {
             entityManager.persist(new FileData(generateHash(file), path));
             byte[] bytes = file.getBytes();
@@ -64,7 +73,6 @@ public class CloudServiceRepository {
         } catch (IOException e) {
             throw new ErrorUploadFileException();
         }
-        //todo добавить обработку, если мы пытаемся загрузить два одинаковых файла.
     }
 
     /**
